+++ src/components/account/beer-info/LikedBeers.tsx
@@ -3,7 +3,7 @@ import {
   useBeerLikeMutation,
 } from "@/../hooks/query/useBeerLikeMutation";
 import { useUserLikedBeersQuery } from "@/../hooks/query/useUserQuery";
-import { BeerResponseType } from "@/../typedef/server/beer";
+import { BeerResponseType } from "../../../../types/server/beer";
 import { generateBeerDetailUrl } from "@/../utils/url";
 import { CommonBeerImage } from "@/components/shared/CommonBeerImage/CommonBeerImage";
 import { LikeButton } from "@/components/shared/LikeButton";
