+++ src/components/search/SearchResult.tsx
@@ -12,7 +12,7 @@ import {
   BeerCountryText,
   BeerNameText,
 } from "../shared/Card/BeerCardItem";
-import { BeerResponseType } from "../../../typedef/server/beer";
+import { BeerResponseType } from "../../../types/server/beer";
 import { generateBeerDetailUrl } from "../../../utils/url";
 import { BeerlotLoading } from "../shared/Loading";
 import { LikeButton } from "../shared/LikeButton";
